Autoscaling/ECS: Setup New Module (#136)
#### Summary
Add autoscaling policy for ECS. 
This module should allow the ECS to autoscale based on the metric threshold + send notification to Slack when it scale in/out.

#### Some thoughts about the decisions
1. Why Step Scaling is chosen instead of Target Tracking
- Target Tracking only limit us to trigger the scaling with limited ECS Metrics
![image](https://user-images.githubusercontent.com/15337955/193999398-28e3d3c7-685c-471c-96e3-6e46aa7119dd.png)
 - Target tracking doesn't give us the flexibility to set the scale down threshold (It automatically create the scale down alarm for us)

2. Why don't we use individual metric alarm?
If we use individual metric alarm, the task count will be increased too quick (> 1) when metrics increase.
Eg. When both memory and cpu utilization is beyond the threshold, it will increase by 2
However, the autoscaling would not be triggered if only 1 metric > threshold.
Eg. When memory is beyond the upper threshold, but cpu is < step down threshold, the increment will be +1 - 1 = 0
And no task count will be added.

3. Why don't we use composite alarm instead of using the metric alarm?
Composite alarm doesn't allow us to use autoscaling as the alarm actions