Fix period override in CloudWatch (#212)
#### Summary

<img width="380" alt="image" src="https://github.com/dbl-works/terraform/assets/20702503/09848ff4-d865-4c65-b2af-ff85449324e2">


#### Motivation

the stats label is "... per min", but if the time range of the dashboard is changed to e.g. 1 week, cloudwatch by default auto-adjusts the period (e.g. to 1 hour), but the label still prints " ... per min", which is very confusing.

This PR fixes this by forcing the dashboard to always respect the configured period regardless of the time range chosen.